Enable SQLite database for testing
The previous version had comments in place of the commands necessary to enable a SQLite database for testing. These comments have been removed and the SQLite database has been activated, setting the connection to SQLite and the database to memory. This change will allow for the use of an in-memory database during testing.