Update dashboard view to display dynamic data
With this change, instead of hardcoding data in the dashboard view, we're actually fetching it from the related ticket object. Now, the priority badge component properly uses the priority of the associated ticket. The email is replaced by the name of the category related to the ticket. Also, the view link now leads to the respective ticket show page.