Update relations in Ticket model
The commit adds HasOne relation to the 'use' statements in the Ticket model. It also updates the `user` function by adding 'owner_id' to specify the relation between the Ticket and User models. Lastly, a new `category` function has been added to describe the relation between the Ticket and Category models.