Add notification feature to reply creation
The commit includes code additions to the CreateReply.php file in the Livewire folder. The new code ensures that whenever a reply is added to a ticket, a notification is triggered to the ticket's owner. A ReplyAdded notification is generated with relevant data and then immediately sent to the user.