Update TicketController with new methods and fields
The TicketController is revamped to include additional methods and fields for better functionality. New methods are added for closing a ticket and retrieving the 'StoreTicketRequest'. The 'store' method is updated with 'StoreTicketRequest' and some fields are modified in the 'store' and 'show' methods. The 'create' method now also fetches all categories.