Implement ticket closing mechanism on tickets view
The code changes contain modifications for the implementation of ticket closing mechanism on the tickets view. This is achieved by adding a POST form action to the 'Close Ticket' button. Furthermore, there are some adjustments to the view to only show the reply creation section and the close button in case the ticket is not already closed. Lastly, a success message is displayed when a ticket has successfully been closed.